common/requirements.txt
# this is a common requirements file that is installed in each service
sonyflake-py
msal
cryptography
PyJWT
networkx
azure-common
azure-core
azure-identity
azure-keyvault-secrets
azure-monitor-opentelemetry
azure-storage-file-share
azure-storage-queue
azure-data-tables
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-core
googleapis-common-protos 
